update create loan with activation channel

+++ fineract-provider/src/main/resources/db/changelog/tenant/parts/0038_m_loan_other.sql
@@ -20,7 +20,7 @@
 CREATE TABLE `m_loan_other`(
       `id` BIGINT PRIMARY KEY AUTO_INCREMENT,
       `loan_id` BIGINT NOT NULL,
-      `activation_channel_id` BIGINT NOT NULL,
+      `activation_channel_id` INT NOT NULL,
       CONSTRAINT `m_loan_other_FK_loan_id` FOREIGN KEY (`loan_id`) REFERENCES `m_loan` (`id`),
       CONSTRAINT `m_loan_other_FK_activation_channel_id` FOREIGN KEY (`activation_channel_id`) REFERENCES `m_code_value` (`id`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_ci;
